Adding worldtree.json
This is the first scan of the worldtree, based on the districts captured by the CRAWL module. The worldtree has all children, reinscriptions, on the tree, and all its inscription info, metadata, and scraped text content. It doesn't yet have parent data, and some limitations on children of reinscriptions which should be fixed in future tree, crawl, & worldtree modules.